+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## unreleased
+### Fixed
+- Fixed canvas size determination
 ### Added
 - Added SiaViewer component (allows viewing annotations without being able to modify them)
 - Added export for getColor utils method (allows other libs to use lost-sias color palette)